Enables feature tests using local or dockerized monolith (#2819)
OKTA-533922 run cucumber feature tests using local (dockerized) monolith

+++ test/e2e/support/management-api/deleteUserAndCredentials.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 /*!
  * Copyright (c) 2015-present, Okta, Inc. and/or its affiliates. All rights reserved.
  * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
@@ -15,8 +16,14 @@ import ActionContext from '../../support/context';
 
 // eslint-disable-next-line no-unused-vars
 export default async function(this: ActionContext): Promise<void> {
+    if (process.env.LOCAL_MONOLITH) {
+        // keep users on local monolith to aid with local debugging
+        return;
+    }
+
+    // remove users in live production org
     if (this.credentials) {
-        await this.a18nClient.deleteProfile(this.credentials.profileId);
+        await this.a18nClient!.deleteProfile(this.credentials.profileId!);
     }
     if (this.user) {
         await deleteUser(this.user);
